debugger: Resolve infinite loop on terminate
I've encountered infinite loop when I've tried to terminate sessions.
Handler of 'terminate' event was generating new 'terminate' event. I've
removed both places where loops was passed on, as it seems it is not needed
in either place.

handleDebugEvents() was used to terminate child processes.
Apparently this should return there in the future, but no need to do
anything with debugSessionEnded() this time.