STAR/9000690333: Debug session ends if vital process has been terminated
Debugging doesn't make sense of OpenOCD or Ashlings GDBserver has been
closed. It also doesn't make sense to keep gdbserver and Putty online if GDB
has exited. This patch adds a monitoring to the state of running processes,
of gdbserver or GDB has terminated, then the rest of the processes will be
terminated as well. Terminating Putty will not have any effect.

This solution uses hard coded process labels and I don't feel secure about
it in the long term, however it seems to work nicely in most cases. If one
of the required processes terminates, then user is presented with a question
window which suggests to end the debug session. If user agrees, then
remaining processes are terminated.

In order for this to work properly I've reordered bits in the
Launch.launch(). So that GDB is created first, then only gdbserver. This is
required to ensure that if gdbserver fails early then GDB still will
terminated. Without this reordering it seems that gdbserver was failing when
GDB was trying to connect to target and Eclipse wasn't able to terminate GDB
at this moment. I'm not completly sure of the reasons of this, but I suspect
that this might be caused by disabling of event handling, however when I was
commenting out this disabling, it still didn't worked as expected, so I'm no
sure.

Signed-off-by: Anton Kolesov <anton.kolesov@synopsys.com>
(cherry picked from commit 1beb9f7cb6a99e661fd9a4b410b3fe2231a47dd6)