Resolve problems with disabling options
Options for HS/EM were not disabled when UI element was created - they were
disabled only when cpu was changed. This patch resolves this, by ensuring that
all setting will be sent to option manager and changing check for processor from
`endsWith()` to `contains()` because sometimes processor options has numeric
suffix in the end.

Signed-off-by: Anton Kolesov <anton.kolesov@synopsys.com>