Fix invalid paths to OpenOCD configuration files
Avoid using INSTALL_DIR dir, instead set path to absolute value when creating
configuration and do not change it automatically after that (like it is done for
the nSIM). Before that basically code was ignoring string in the Debug
configuration if it contained INSTALL_DIR and was replacing it with a default
value.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>