Fix Run As and do not stop at __start initially
This patch fixes two problems.

First, "Run" mode behaviour was wrong as it was ignoring when target was
actually hitting "flag 1" in the _exit_halt(). Now Eclipse will understand that
target stops in this location.

Secondly, previously debugging was always started at __start, next "resume" was
stopping executing at main() (if this feature was enabled). After this patch
application will always skip __start and will go straight to main(). If stopping
at main() is disabled in configuration, that it will go on, until first
breakpoint hit.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>