9000753695: Update crt0.s in templates
Do not clobber register R0 when handling exception in default crt0.s file of
our project templates. Same function _exit_halt() is used to implement
exit() function and to handle exception. In the former case we want R0 to
contain exit code of application, but in the latter case we want all
registers to be left untouched. Previously our exception handlers would
overwrite R0 with non-zero value similar to ECR. This code is removed now,
so instead user has to first check value of STATUS32.AE bit. If it is 0,
then R0 contains exit code, if it is not, then it contains genuine R0 value,
and user has to look at ECR register.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>