9000815947: Fix bug with invalid handling of virtual paths
We used to treat any file path as a physical path, but Eclipse allows
creation of it's own virtual folders and links (not sure why they just don't
use native OS capabilities for this and invent a bicycle). That was causing
an issue as we were searching for file, that doesn't exists physically. This
commit adds an additional step to the path handling algorithm and now it
will try to resolve file path as a virtual/linked path.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>