Fix and refator detection of tools
Originally there was a bug - Eclipse always tries to use tools from
predefined path ("../bin" relatively to the Eclipse) without looking
into PATH environment variable. I've introduced a method
CommandInfo.commandExistsInPredefinedPath which helps to check whether
tool exists in the predefined path or not. IsToolChainSupported uses
this method to decide whether to append the predefined path to to the
command or not.

Also refactoring and decomposition of CommandInfo class is performed.

Signed-off-by: Yuriy Kolerov <yuriy.kolerov@synopsys.com>