9000922449 Detect whether Linux toolchain is applicable for ARC 700 or HS
It's implemented by scanning output of "gcc -v" command and looking
for "--with-cpu=archs" substring. Than it's necessary to check does
toolchain's name contain "ARC HS" substring and make a decision
about enableing of the Linux toolchain.

The only problem is that this algorithm demands tollchain name to
contain a substring like "ARC 700" or "ARC HS". Thus it's necessary
to be careful in the future.

Signed-off-by: Yuriy Kolerov <yuriy.kolerov@synopsys.com>
Signed-off-by: Yunlu Zhang <Yunlu.Zhang@synopsys.com>