Fix storing JTAG frequency field value
If in IDE we applied Opella-XD debug configuration without changing
default value of JTAG frequency, it wouldn't be passed to back end
configuration and therefore GDB server.
Another thing was that if we already had applied some value of JTAG
frequency, instead of it in IDE default value was showed in some cases.

Now the default value is passed to back end as well as other ones.
Also if there is a value of JTAG frequency applied earlier it is
selected when we open Debugger tab but not default one.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>