Fix problem with building big endian projects
There was a problem in the case we had only big endian toolchain
available in the path. Then we should have been able to build Hello
World project corresponding to that toolchain, but we couldn't.

The reason was because we had ScannerConfigurationDiscoveryProfiles only
for little endian case and their specs providers called little endian
tools.
This commit adds ScannerConfigurationDiscoveryProfiles for big endian
case to both elf32 and uclibc plugin.xml files.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>