Add sleep after launching Ashling GDB Server
This sleep is required so that Ashling server has time to start before
GDB connects to it. Otherwise, if GDB tries to connect to a server that
hasn't started listening yet, we get an exception and GDB terminates.

The other thing we might do to make sure Ashling has started is to wait
for a specific message Ashling writes to output stream. However, it
seems impossible to do it in our case because for some reason Ashling
doesn't write any messages to neither output nor error stream when
everything goes as supposed. It probably happens because event listener
is added to the process after all the messages are already sent. Then
the listener isn't called until Ashling send something else to the
stream, which doesn't happen in our case.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>