9000835886 Support TCF as an input option in IDE
Added TCF button to Target Options tab for EM and HS toolchains.
If it is selected, all other architecture options get disabled and their
values don't go to the compiler.

It is possible to use not only compiler options from TCF, but also pass
memory map from TCF to linker. Supported this option in IDE as well.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>