Add check that selected TCF file is valid
Check that if 'TCF' button is selected, value of 'TCF path' field is not
empty and that selected TCF exists. Also check that this file is valid
TCF, it has 'gcc_compiler' configuration and -mcpu option there
corresponds to that of the current tool chain.

If any of these checks haven't passed, show error dialog.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>