Escape whitespace characters when reading TCF
Compiler options from TCF are read using java.util.Properties, which
consideres '=', ':' and whitespace characters as key termination
characters. So if our TCF contains options like '--param ', --param is
considered a key here and therefore only one of these options can be
read.

But if a whitespace is escaped, it is seen as part of the key and all
the --param options can be read.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>