9000985736: Fix invalid syntax error for big endian
On Windows Eclipse wasn't able to find location of standard library headers
for big endian toolchain. As a result headers like "stdio.h" were marked as
"unknown" and thus all of the functions, like printf were unresolved as
well. This problem was only with Eclipse syntax checker, application was
built without an error.

It was found that this happens because for big endian Eclipse tries to
figure out location of stdlib headers via invocation of `gcc` instead of
`arceb-elf32-gcc`. Hence the problem itself was not specific to Windows, it
is just that on Linux systems we usually have `gcc` in the PATH, so headers
of host system were used for syntax check. Everything was find for little
endian systems.

It was notoriously hard to understand where the difference between little and
big endian comes from, in part because big endian toolchain inherits from
little endian one, so XML definitions for them are different per definition.
Investigation became easier after it was realized that this difference
happens only for compiler, but not for assembler, and after comparing big
endian assembler and compiler it was found that redefinition of `<inputType>`
for big endian compiler is not needed. It is not clear to me how this
additional `<inputType>` was causing the behaviour we were observing, but
nevertheless problem was fixed after I removed this extra tag.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>