Support 'import' of options from TCF in IDE
If TCF is selected, then all target options that are present in both
TCF and Target Processor tab, are in IDE set equal to the ones in TCF.
If an architecture option is present only in Target Processor tab, then
it is set to its default value. If default value isn't specified, it is
considered false for options with boolean value type and empty string
for options with string value type. All the other options in IDE are
left unchanged.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>