Separate RXTX library from Debugger plugin
Debugger plugin depends on TM Terminal plugin, which requires RXTX
library, so we need to include this library in our plugin. But in the
case when Debugger contains RXTX there are circular dependencies
(Debugger depends on TM Terminal and TM Terminal depends on Debugger)
and our plugin can not be built by ant because of this.

To avoid this problem RXTX library was moved from Debugger plugin into a
separate one.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>