9000970301 Create ant scripts for automatic build
There are individual ant scripts for each plugin (Debugger, Elf32,
Uclibc and newly added RXTX) and one ant script for feature.

These scripts were generated by Eclipse PDE, but they contained errors
and were not usable as they were, so they were modified manually for our
purposes. Ant script for a plugin should be updated every time a
dependency is added for this plugin. If plugin project contains
javaCompiler.<library>.args file, it also should be updated: a line for
this new plugin dependency should be added similar to the existing ones.
In brackets after library name there are packages that are visible to
the compiler. If compiler should see all the packages, use [?**/*].

In order to build the project, makefile in build/ directory should be
used. It creates arc_gnu_plugins/ directory and arc_gnu_plugins.zip
archive, both of which can be used as update sites, from which plugin
can be installed into Eclipse.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>