Update -mcpu dropdown lists
Create values in dropdown lists for new values of -mcpu option for EM
and HS and change default values of other target options so they would
be turned off by default.

It is necessary because when "Restore Defaults" button is pressed, all
the options' default values are set from the beginning of the page to
the end. So if default values of options were not the same as their
values for default CPU, they would be overwriten. Since default values
of CPU are -mcpu=em and -mcpu=hs where all the options are turned off,
default values should also be "turned off".

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>