Change Properties to OrderedProperties
Properties were used to store gcc options before, but Properties do not
keep the order of the options, so they are not suitable for this
purpose. Options order is important because if several values are
specified for the same option, only the last one should be set.
For example, it is important if we set -mcpu value first and then
-mmpy-option value or the other way around since -mcpu option implicitly
sets multiplier value and which of these values will be used depends on
their order.

Created OrderedProperties to use instead of Properties for storing gcc
options to avoid problems described above.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>