Disable checkboxes set because of -mcpu value
For each CPU value there is a standard library which uses certain
options. These options are set in IDE accordingly, and checkboxes are
disabled so that user could not unselect them. If he unselected one of
the set for CPU checkboxes, he might think that this option was not used
anymore, but in fact it still would be used by standard library.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>