Show error messages if inapplicable value chosen
For option values that are presented in dropdown lists user should not
be able to choose value that is weaker than the one set by CPU for the
same reason as he should not be able to uncheck checkboxes. But he
should be able to select stronger values, so we do not disable dropdown
lists. It is not possible to disable only part of the values in dropdown
lists, so we show error messages if user tries to select weaker value
and replace the value he chose with default one for selected CPU.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>