Store options values to be set in a map
Instead of setting options in gui right away should use map to store
options values. It is necessary when read option values from TCF we at
first set options according to CPU value and then possibly rewrite their
values with what is written in TCF. If weset these options in gui like
it was done before, events about options values changing come in some
random order, so in the end values can be set not to the last value we
set but to some other one.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>