Check architecture for TcfContent from cache
When we read TcfContent, we check that TCF is for the same toolchain as
used in our project. But if TcfContent is not read but obtained from
cache, there was no toolchain check. So if TCF was read for one
toolchain and then also selected for another, check wouldn't be done.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>