Read mcpu option value from TCF when TCF is chosen
Before mcpu value was read from toolchain configuration and there was no
difference because for toolchain there was only one valid mcpu value and
there was a check that project and TCF are for the same toolchain. So
you could not use TCF with cpu value different from toolchain's.

Now there are several valid CPU values for toolchain, so we have to read
mcpu option value from TCF as well as other target options' values.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>