Add error for bad working directory on nSIM start
It is allowed to save incorrect values of working directory and other
variables to configuration. In this case 'Debug' button in Debug
configuration dialog becomes disabled, but user still can launch this
debug configuration by pressing IDE debug button if this configuration
is in the list of recent configurations.

In this case, if working directory is incorrect, it is replaced with
system user.dir property. This commit adds showing error message if
user.dir is used instead of incorrect working directory.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>