Search for toolchain first in PATH, then in ../bin
If toolchain for a project is present in both PATH and ../bin directory,
toolchain from ../bin was chosen when creating a new project. But specs
provider in this case was used from toolchain contained in PATH.

To make toolchain choosing consistent, now toolchain from PATH is
selected in both these cases.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>