9001027655 Build both C and C++ files in a project
Before for C++ projects only G++ compiler was available, so C files
could not be compiled. Similarly for C projects it was impossible to
compile C++ files. This commit adds GCC and G++ to both C and C++
projects so that C files were compiled by GCC and C++ files by G++
regardless of project type.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>