9000968722 Migrate from TM Terminal 3.7 to 4.0
Our debugger plugin uses Target Management Terminal plugin for
connecting to the UART port of our boards. We used version 3.7 of this
plugin before, which had been deprecated for some time, so it was
decided to migrate to version 4.0 from it.

From 3.7 to 4.0 there were a lot of changes in TM, and the class
TerminalView we were using before for opening a connection got split.
Therefore the code that was connecting to UART and opening a terminal
view needed to be completely rewritten.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>