Add utility class to get launch config options
To get an option value from launch configuration you had to call
launchConfiguration.getAttribute() method often surrounded by try/catch.
Also you had to specify the key under which the option value was stored
every time, which wasn't convenient since you had to find out which key
to use. Another thing about these keys is that if the key had changed,
you had to change it in several places in the code.

Now it can be done by calling one of the methods in Configuration class.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>