Use CMainTab2 instead of CMainTab
On CMainTab the first field is 'C/C++ Application' and only after that
'Project', but their values have to specified in reverse order.
CMainTab2 is more consistent since these fields there are in the same
order as they have to be filled.

Also CMainTab2 is used now for standard CDT 'C/C++ Application' debug
configurations, so it's better to have the same tab as they do than
the different one.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>