Change Debugger main tab in debug configurations
Use Debugger tab from DSF UI package. This way the options from this
page are saved into launch configuration using DSF constants as keys.
When launching a configuration using DSF there options would be read
using the same keys and passed to debugger by DSF automatically.

Otherwise, if we left the page as it was and therefore used the old keys
to store the options in configuration, we would have to read the options
ourselves and pass them to debugger.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>