Stop debug session if GDB server process exited
We use standart GDBBackend from DSF which monitors the GDB process and
stops the debug session in case it terminated.

However, GDB server backend does not have a way to track if GDB server
process is terminated or not, so we need to listen to the process
termination events and stop the debug session manually.

Signed-off-by: Anna Pologova <anna.pologova@synopsys.com>