Add ConfigurationReader.java
This class is not used in code yet, so it does not break anything.
ConfigurationReader.java will be the replacement of Configuration.java,
the enhancements are listed below.

Replace static methods with non-static, store ILaunchConfiguration
in a field, there is no need to pass it in every method now.

Replace printing a stacktrace when the CoreException occurs with the
StatusManager's handle method call, it is a preferred way of handling
this exception (the current status with the stacktrace will be logged).

Remove parsing boolean from string in the get methods, it is better to store
boolean (will be done in one of the next commits) and retrieve it (change
the return type of get methods to boolean, done in this commit).

Add "Locale.ENGLISH" as an argument to "toLowerCase()" in
"getAshlingPath()" to avoid local issues, e.g. related with lowercasing
the character "I" in Turkey (it becomes lower case undotted "i" ("ı")).

Add "getNsimUseDefaultDirectory" method, because the corresponding boolean
attribute is retrieved in the "ARCWorkingDirectoryBlock.java" manually.

Replace "(String) null" in the method "getNsimWorkingDirectoryPath" with an
empty string, because there is no need to differ this method from the another
get methods returning a string.

Add the IllegalArgumentException handling to "getFtdiDevice(), getFtdiCore()"
when calling the enum's "valueOf()".