uClibc: Recognize GCCs for hs38*
uClibc plugin checks the GCC configure string to identify if compiler was
configured to build for ARC HS or for ARC700.  This stopped working properly
with recent releases of toolchain, because they are now configured by
default for -mcpu=hs38 while it was -mcpu=archs in the past.  This patch
solves this issue: uClibc plugin now checks for both `archs` and `hs38`.
Pattern for `hs38` will also recognize `hs38_linux` which is a valid CPU for
Linux as well.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>