Fix bug with custom GDBserver argument UI element not shown
There was a bug introduced in a commit 36edf71a "Extract
customGdbCommandLineArguments and customGdbCommandLineArgumentsText from
RemoteGdbDebuggerPage" - code to create an UI element to input server
arguments was extracted to a separate function, but this function was never
invoked. This patch fixes this bug by adding invocation where necessary. It
also makes this function private instead of public, since it is never used
outside of this class.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>