Add support for generic EM SK v2.x
Add a separate entry on debugger plugin to support any EM Starter Kit v2.x,
instead of a separate entries for 2.1 and 2.2. This is done to seamlessly
support v2.3. Discrete entries for 2.1 and 2.2 are retained for
compatibility with existing projects, but might be removed in the next
release to avoid confusion among users.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>