Add a fixed length artificial delay after starting GDB server
Add a delay after starting GDBserver process, so that it would have time to
start listening on TCP port before client will attempt to connect to it.
Without this delay there is a risk that GDB will try to connect before
server starts to listen so GDB will exit prematurely and the whole session
will be aborted.

This solution is far from perfect because it uses same delay for all cases,
so it might not help in some cases where start is even slower - it would be
better to at least make this user configurable, or even better to scan
server output for a line that notifies that it is listening. Then that could
be handled via some asynchronous DSF magic, I presume. However right we
don't have time for fancy solutions, and this gets it working at least in
most cases.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>