Ensure compatibility between manifest and Maven output
By default, Eclipse puts compiled classes files into a "bin/" folder in the
generated JAR files. However Maven instead puts them directly into the root
of the JAR Files and as a result at runtime Eclipse cannot find those
classes. To fix this it is needed to change manifest file to specify proper
location of the compiled classes. In development Eclipse would keep using
"bin/" folders but it seems that even if plugin is exported via Eclipse it
still would be compatible with Maven setup.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>