Fix issue with debug configuration reset
There was found a bug with Eclipse Neon, that changes to ARC debugger
configuration were not permanent - each time user would select ARC
debugger configuration it would be reset to default values. Eventually
the source of this trouble has been traced to
EmbeddedDebuggerTab2.initDebuggerTypes which indeed sets a field that
specifies that configuration should be set to default value. This seems
unnecessary, because new configuration is already set to defaults
earlier, and for an existing configuration that just resets
configuration.

Because this code has been developed not by us it is not clear why this
`setInitializeDefault(true)` was done there in the first place - perhaps
it makes sense in some cases?

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>