9001248451: Show gdbserver output in Eclipse console
Eclipse Neon stopped directing gdbserver output to the Console. As a
result it became impossible to observe output of Hello World running on
nSIM, or looking at OpenOCD output, etc.

It appeared that problem happens because gdbserver server was started
as a "GDB"-type process, and Eclipse started to hide output for this
kind of processes. If started as a "default"-type process output is
present as it used to be.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>