debug: Change icon size to 16x16 pixels
We used to have an ARC image of size 32x32 pixels and Eclipse was
stretching rest of GUI to be bigger than the default. By reducing size
of image to the standard 16x16 GUI now looks as in default Eclipse.

We'd need a better image (content is unintelligible at this size) but
that is a task for another day.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>