debug: Add support for adaptive server startup delay
This is a better version of commit

2775a84 Add a fixed length artificial delay after starting GDB server

When adaptive delay is used, debugger plugin will try to connect to the
server several times, until specified time runs out or server is
listening. This makes sure that GDB will try to connect to the server
only after it is ready for connection. Main driving problem is that of
nSIM is used with some remote license server it may take up to several
seconds before it starts listeining for incomming connections. Adding a
fixed delay of that size is not good, because this delay would be
applied in all cases, even when it is not needed, as a result it would
force everyone to wait for significant time. To avoid cases where
Eclipse will poll server indefinetely, there is a up limit on how much
time can be spent at this stage, this limit can be set through
plugin preferences.

One problem with current implementation is that even if this algorithm
would give up, it still would try to run GDB client to connect to the
server, which will fail, of course, and will trigger a rather cryptic
error message. So this fixes a "good" case where server does eventually
starts, but it doesn't improve overall usability situation for error
cases.

Note that this algorithms cann't be used in all cases. For example, but
default nSIM would immediately close after client disconnects, hence it
would close after the probe connection. Thus nSIM should be started with
`-reconnect` option (which it isn't right now). OpenOCD and Ashling
should work fine out of the box. It is impossible  to predict situation
for external GDBservers, so for now this algorithm shall not be applied
to them (at least until a GUI configure option will be added, that will
allow user to make choice themselves).

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>