repo: Add missing qualifier to feature version
Specification of a feature version in repository/category.xml was
without a version qualifier, which used to work fine, until commit [1]
which removed from feature/pom.xml custom version without a qualifier
and switch to a standard version from parent pom.xml. As a result of
this change repository failed to use feature properly and output
directory didn't contain it.

[1] 30705c6 maven: Remove duplicate artifact version from pom.xml files