elf: Remove extra configuration elements
Experiments has shown that it is enough to have a single base
configuration element that can be reused by various configurations that
have to exist in each project type. This commit removes all
configuration elements except for one, some of those configurations
already were not used anywhere by this commit.