elf: Allow SCDP Provider to find compiler at predefined path
This commit modifies ARC GCC Specs files provider for SCD profile. The
provider should be able to find GCC at the predefined location of "bin"
folder relatively to Eclipse, even when it is not in the PATH.
Previously that was achieved by appending bin to the PATH dynamically,
but that can be also achieved by recalculating the path correctly during
the initialization, which this commit does.