gh#265: Don't show warning for new project version mismatch
This commit changes behaviour of when warning for the mismatch of
project version is shown. Previously it would be shown for a "new"
project if it is not in the workspace, even if version is not set -
assuming that any project not in workspace is an import of existing
project. That assumption has been proven to be untrue, because a new
project from existing sources looks exactly the same, and we shouldn't
throw a warning for it. Unfortunately I don't know how to distinguish
"imported project" from "imported sources", so now plugin will not show
a warning for any project without an explicit version.

In the process I've noted that overall this code doesn't work, because
even if imported folder contains a .settings folder with some value for
project version, then this preference is ignored and project gets a new
version now. This is true even when importing an existing project.
Overall that code with version checking is not something I like, and it
seems like it doesn't really work as expected.

Fixes https://github.com/foss-for-synopsys-dwc-arc-processors/toolchain/issues/265